+++ packages/dara-core/dara/core/defaults.py
@@ -21,6 +21,8 @@ from typing import TYPE_CHECKING, cast
 
 from dara.core.base_definitions import ActionDef
 from dara.core.interactivity.actions import (
+    CopyToClipboard,
+    CopyToClipboardDef,
     DownloadContentDef,
     DownloadVariable,
     DownloadVariableDef,
@@ -102,6 +104,7 @@ CORE_ACTIONS: dict[str, ActionDef] = {
     TriggerVariable.__name__: TriggerVariableDef,
     ResetVariables.__name__: ResetVariablesDef,
     DownloadVariable.__name__: DownloadVariableDef,
+    CopyToClipboard.__name__: CopyToClipboardDef,
     'DownloadContent': DownloadContentDef,
     Notify.__name__: NotifyDef,
 }