+++ packages/dara-core/js/actions/index.tsx
@@ -4,3 +4,4 @@ export { default as TriggerVariable } from './trigger-variable';
 export { default as UpdateVariable } from './update-variable';
 export { default as DownloadVariable } from './download-variable';
 export { default as Notify } from './notify';
+export { default as CopyToClipboard } from './copy-to-clipboard';