+++ packages/dara-core/changelog.md
@@ -2,6 +2,10 @@
 title: Changelog
 ---
 
+## NEXT
+
+-  Added a new `CopyToClipboard` action (also available as `ctx.copy_to_clipboard`) to copy a given value to the user's clipboard
+
 ## 1.21.22
 
 -  Allow passing a `ClientVariable` to the `DynamicComponent` component